+++ front/src/app/modules/utils/utils.module.ts
@@ -2,17 +2,20 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { TruncatePipe } from './truncate.pipe';
 import { GenerateJoinLinkPipe } from "./generate-join-link.pipe";
+import { GenerateMagicLinkPipe } from './generate-magic-link.pipe';
 
 
 
 @NgModule({
   declarations: [
     TruncatePipe,
-    GenerateJoinLinkPipe
+    GenerateJoinLinkPipe,
+    GenerateMagicLinkPipe
   ],
   exports: [
     TruncatePipe,
-    GenerateJoinLinkPipe
+    GenerateJoinLinkPipe,
+    GenerateMagicLinkPipe
   ],
   imports: [
     CommonModule